Add sqlite_health.c/h
