+++ CMakeLists.txt
@@ -619,6 +619,8 @@ set(RRD_PLUGIN_FILES
         database/sqlite/sqlite_aclk_alert.h
         database/sqlite/sqlite_aclk_node.c
         database/sqlite/sqlite_aclk_node.h
+        database/sqlite/sqlite_health.c
+        database/sqlite/sqlite_health.h
         database/sqlite/sqlite3.c
         database/sqlite/sqlite3.h
         database/engine/rrdengine.c
