+++ database/sqlite/sqlite_functions.h
@@ -82,14 +82,15 @@ extern void compute_chart_hash(RRDSET *st);
 extern void sql_chart_from_hash_id(char *hash_str, BUFFER *wb);
 extern int alert_hash_and_store_config(uuid_t hash_id, struct alert_config *cfg);
 extern void sql_select_alert_config(char *hash_str, BUFFER *wb);
-extern void sql_health_alarm_log_load(RRDHOST *host);
-extern int sql_create_health_log_table(RRDHOST *host);
-extern void sql_health_alarm_log_update(RRDHOST *host, ALARM_ENTRY *ae);
-extern void sql_health_alarm_log_insert(RRDHOST *host, ALARM_ENTRY *ae);
-extern void sql_health_alarm_log_save(RRDHOST *host, ALARM_ENTRY *ae);
-extern void sql_health_alarm_log_select_all(BUFFER *wb, RRDHOST *host);
-extern void sql_health_alarm_log_cleanup(RRDHOST *host);
+//extern void sql_health_alarm_log_load(RRDHOST *host);
+//extern int sql_create_health_log_table(RRDHOST *host);
+//extern void sql_health_alarm_log_update(RRDHOST *host, ALARM_ENTRY *ae);
+//extern void sql_health_alarm_log_insert(RRDHOST *host, ALARM_ENTRY *ae);
+//extern void sql_health_alarm_log_save(RRDHOST *host, ALARM_ENTRY *ae);
+//extern void sql_health_alarm_log_select_all(BUFFER *wb, RRDHOST *host);
+//extern void sql_health_alarm_log_cleanup(RRDHOST *host);
 extern int execute_insert(sqlite3_stmt *res);
 extern void compute_chart_hash(RRDSET *st);
 extern void sql_chart_from_hash_id(char *hash_str, BUFFER *wb);
+extern int prepare_statement(sqlite3 *database, char *query, sqlite3_stmt **statement);
 #endif //NETDATA_SQLITE_FUNCTIONS_H
